+++ modules/components/widgets/antd/Select.jsx
@@ -1,7 +1,6 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import map from 'lodash/map';
-import {useOnPropsChanged, calcTextWidth, SELECT_WIDTH_OFFSET_RIGHT} from '../../../utils/stuff';
+import {useOnPropsChanged, mapListValues, calcTextWidth, SELECT_WIDTH_OFFSET_RIGHT} from '../../../utils/stuff';
 import { Select } from 'antd';
 const Option = Select.Option;
 
@@ -14,7 +13,7 @@ export default class SelectWidget extends PureComponent {
     customProps: PropTypes.object,
     fieldDefinition: PropTypes.object,
     // from fieldSettings:
-    listValues: PropTypes.object,
+    listValues: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
   };
 
   constructor(props) {
@@ -27,13 +26,13 @@ export default class SelectWidget extends PureComponent {
     const {listValues} = props;
 
     let optionsMaxWidth = 0;
-    map(listValues, (label, value) => {
-      optionsMaxWidth = Math.max(optionsMaxWidth, calcTextWidth(label));
+    mapListValues(listValues, ({title, value}) => {
+      optionsMaxWidth = Math.max(optionsMaxWidth, calcTextWidth(title, null));
     });
     this.optionsMaxWidth = optionsMaxWidth;
 
-    this.options = map(listValues, (label, value) => {
-      return (<Option key={value} value={value}>{label}</Option>);
+    this.options = mapListValues(listValues, ({title, value}) => {
+      return (<Option key={value} value={value}>{title}</Option>);
     });
   }